+++ resources/views/livewire/summary.blade.php
@@ -59,8 +59,8 @@
                                     <a href="#" wire:click.prevent="editAnswer({{ $question->node_id }})"
                                        class="nhsuk-link nhsuk-task-list__link">
                                         <strong>{!! $question->title !!}</strong>
+                                        <span class="nhsuk-u-visually-hidden">Edit this answer</span>
                                     </a>
-                                    <span class="nhsuk-u-visually-hidden">Click to edit this answer</span>
                                 @endif
 
                                 <br>